# Faster `minimum`/`maximum`/`extrema` for `Float64/32`