@oscardssmith 2022-01-09
Has this been sufficiently `NaN`/`Inf` tested?