@N5N3 2022-01-09
`NaN`, `Inf`, `-0.0,0.0` shoud have been tested (previously) in #43604.
As for `missing`, I limited this to concrete `IEEEFloat`. Thus inputs with `missing` will use general fallback.