@oscardssmith 2022-01-09
It would be cool to get simd for unions with singletons (which should be able to be done pretty well). Doesn't have to be in this PR though.