@N5N3 2022-01-09
I make some try to make `Union{Missing}` vectorlizable. (The unroll size is limited to `4`).
Benchmark:
```julia
for T in (Float32, Float64, Int32, Int64)
    a = Vector{Union{Missing,T}}(rand(T, 4096))
    print(T,":")
    @btime Base.mapreduce_impl(identity, min, $a, 1, 4096)     
end
```
On master, general fallback:
```julia
Float32:  11.700 μs (0 allocations: 0 bytes)
Float64:  11.100 μs (0 allocations: 0 bytes)
Int32:  6.580 μs (0 allocations: 0 bytes)
Int64:  6.600 μs (0 allocations: 0 bytes)
```
On master, optimized fallback:
```julia
Float32:  1.970 μs (3 allocations: 64 bytes) # Don't know where cause these allocations
Float64:  1.930 μs (3 allocations: 64 bytes)
Int32:  1.610 μs (4 allocations: 96 bytes)
Int64:  1.740 μs (4 allocations: 96 bytes)
```
On 1.7.1
```julia
Float32:  8.300 μs (0 allocations: 0 bytes)
Float64:  8.000 μs (0 allocations: 0 bytes)
Int32:  7.300 μs (0 allocations: 0 bytes)
Int64:  7.325 μs (0 allocations: 0 bytes)
```
The performance for non `missing` inputs does not change.