@N5N3 2022-01-09
`sum` is also not optimized for `Union`.
So the better choice to me is to implement a general `mapreduce_impl` for `Union{Missing,T}` with a fixed unroll size of 4, which should also accelerate other reduction.