@N5N3 2022-01-19
`extrema` related optimization has been added, the latest benchmark for `Float64`:
```julia
2:extrema:9.6ns minimum:5.0ns
4:extrema:14.3ns minimum:7.8ns
8:extrema:24.2ns minimum:13.2ns
16:extrema:33.9ns minimum:18.6ns
32:extrema:37.4ns minimum:20.6ns
64:extrema:45.3ns minimum:24.1ns
128:extrema:55.3ns minimum:33.0ns
256:extrema:74.9ns minimum:47.2ns
512:extrema:115.2ns minimum:75.6ns
1024:extrema:195.6ns minimum:132.2ns
2048:extrema:358.3ns minimum:247.3ns
4096:extrema:710.5ns minimum:492.3ns
```