@N5N3 2022-01-28
Since #43573 has not been landed, I tried to use `Ref{NTupe{16}}` to emulate `MVector{16}`.
Local bench shows it works well. (TTFP is reduced about a half, and runtime performance is not impacted.)